requirements/common.txt
# Please update the date when you decide a package's version
# NOT_LATEST means a newer is version available, but it was decided not to update
# UNMAINTAINED means the package owner is no longer maintaining it

# Core
Django>=3.1,<3.3  # ------------------------------------------------------------ (bumped 2021-06-24)
django-environ>=0.4,<0.11
sqlparse>=0.2,<0.5  # (required for some old migrations) ----------------------- (bumped 2021-04-15)

# Django Utils
django-fsm>=2.7,<2.8  # -------------------------------------------------------- (bumped 2021-04-15)
django-filter>=2.4,<2.5  # ----------------------------------------------------- (bumped 2021-04-15)
django-livefield>=3.3,<3.4  # -------------------------------------------------- (bumped 2021-04-15)
django-model-utils>=4.1,<4.2  # ------------------------------------------------ (bumped 2021-04-15)
django-annoying>=0.10,<0.11  # (various Django helpers) ----------------------- (checked 2021-04-15)
django-autocomplete-light>=3.9,<3.12  # ---------------------------------------- (bumped 2024-08-07)

# API
djangorestframework>=3.12,<3.13  # --------------------------------------------- (bumped 2021-04-15)
djangorestframework-bulk<0.3  # ----------------------------------------------- (checked 2021-04-15)

# I18n
pycountry>=20.7.3  # ----------------------------------------------------------- (bumped 2021-04-15)
python-dateutil>=2.8,<2.9  # --------------------------------------------------- (bumped 2021-04-15)
pyvat>=1.3,<1.4  # ------------------------------------------------------------ (checked 2021-04-15)

# Crypto
cryptography>=3.3,<3.4  # ------------------------------------------------------ (bumped 2021-04-15)
PyJWT>=2.0,<2.1  # ------------------------------------------------------------- (bumped 2021-04-15)

# Other
furl>=1.2,<1.3  # (URL parsing and manipulation) NOT_LATEST -------------------- (bumped 2018-08-10)
xhtml2pdf>=0.2,<0.3  # (PDF rendering, python-dev is required) ----------------- (bumped 2018-06-07)
PyPDF2>=1.26,<2  # (PDF manipulation) ------------------------------------------ (bumped 2018-06-07)
psycopg2-binary>=2.9,<3

#Celery
celery[redis]==5.3.*
celery-once==3.0.1
